requirements.txt
# UI
customtkinter>=5.2.0
Pillow>=10.0.0

# Kéo thả file audio trên Windows (tùy chọn, bonus)
windnd>=1.0.0; sys_platform == 'win32'

# Tách nền avatar — cài khi bật checkbox rembg (nặng, tải model lần đầu)
# rembg>=2.0.50
# onnxruntime>=1.16.0
